#101 #94 Finally go the contract typeing to work.  Now i need to work on the reducer
