+++ src/context/index.ts
@@ -1,4 +1,4 @@
 export * from './ethers-app/BlockNumberContext';
 export * from './ethers-app/EthersAppContext';
 export * from './ethers-app/connectors';
-export * from './contracts-app/contractContextFactory';
+export * from './contracts-app';
